+++ R/sample_frac.R
@@ -1,12 +1,7 @@
 #' Sample n rows from a disk.frame
-#' @param tbl a disk.frame
-#' @param size the proportion/fraction of rows to sample
-#' @param replace TRUE to sample with replacement; FALSE to sample without replacement
-#' @param weight weight of each row. NOT implemented
-#' @param .env for compatibility
-#' @param ... passed to dplyr
 #' @export
 #' @importFrom dplyr sample_frac
+#' @inheritParams dplyr::sample_frac
 #' @rdname sample
 #' @examples
 #' cars.df = as.disk.frame(cars)
@@ -15,11 +10,18 @@
 #' 
 #' # clean up cars.df
 #' delete(cars.df)
-sample_frac.disk.frame <- function(tbl, size = 1, replace = FALSE, weight = NULL, .env = NULL, ...) {
+sample_frac.disk.frame <- function(tbl, size=1, replace=FALSE, weight=NULL, .env=NULL, ...) {
+  warning_msg = NULL
   if(!is.null(weight)) {
-    stop("sample_frac(..., weight =) is not implemented yet")
+    warning_msg = "sample_frac: for disk.frames weight = is not supported"
+    stop(warning_msg)
+  } else if (!is.null(.env)) {
+    warning_msg = "sample_frac: for disk.frames .env = is not supported"
+    stop(warning_msg)
   }
   
-  delayed(tbl, ~sample_frac(.x, size, replace, weight, .env, ...))
+  fn = disk.frame::create_dplyr_mapper(dplyr::sample_frac)
+  
+  fn(tbl, size = size, replace = replace, ...)
 }