+++ NEWS.md
@@ -1,3 +1,10 @@
+# disk.frame 0.2.1
+* deprecated group_by, arrange, summarise
+* add chunk_group_by, chunk_arrange, chunk_summarise
+* fit GLMs with `dfglm`
+* fixed so that dplyr function also work in mutate even with ~ in the name
+* fixed disk.frame so that in works in functions too
+
 # disk.frame 0.1.1
 
 * Allowed `map` to accept multiple arguments. Thanks Knut Jägersberg for suggestion